Move AccessDenied from src/components to src/views
