+++ app/javascript/src/views/Interview/index.js
@@ -1,7 +1,9 @@
 import React from "react";
 import { Switch, useParams } from "react-router-dom";
+import AccessDenied, { isNotAuthorized } from "src/views/AccessDenied";
+import NotFound, { isNotFound } from "src/views/NotFound";
 import Route from "src/components/Route";
-import Loading from "components/Loading";
+import Loading from "src/components/Loading";
 import { useInterview } from "./queries";
 import CallScheduled from "./CallScheduled";
 import CallRequested from "./CallRequested";
@@ -10,8 +12,6 @@ import MoreTimeOptionsAdded from "./MoreTimeOptionsAdded";
 import ClientRequestedReschedule from "./ClientRequestedReschedule";
 import SpecialistRequestedReschedule from "./SpecialistRequestedReschedule";
 import RescheduleInterview from "./RescheduleInterview";
-import NotFound, { isNotAuthorized, isNotFound } from "../NotFound";
-import AccessDenied from "src/components/AccessDenied";
 
 function InterviewState({ interview }) {
   switch (interview.status) {
