+++ app/javascript/src/views/Project/index.js
@@ -2,12 +2,12 @@ import React, { lazy } from "react";
 import { useQuery } from "@apollo/client";
 import { Container, useBreakpoint } from "@advisable/donut";
 import { useParams, useLocation, Switch, Redirect } from "react-router-dom";
+import AccessDenied from "src/views/AccessDenied";
 import Route from "src/components/Route";
 import View from "components/View";
 import Loading from "components/Loading";
 import { handleAuthError } from "./utilities";
 import NotFound, { isNotFound } from "../NotFound";
-import AccessDenied from "components/AccessDenied";
 import { GET_PROJECT } from "./queries";
 import ProjectRoutes from "./ProjectRoutes";
 import Navigation from "./Navigation";
